Update optimal.py (#3)
There is a bug in how optimal symbolic shift polynomials are computed when M_bar is strictly larger than M. After computing the echelon form of S_bar, we must select basis vectors whose coefficients are zero in M_bar \ M. The bug led to the wrong columns being checked. The echelon form is not used in the non-symbolic code, so the bug only needs to be fixed here.